Move all the configs to a config.js file & import multer & cloudinary for files upload
