+++ backend/.env
@@ -1,3 +1,4 @@
 PORT =
 MONGODB = 
 SECRET_KEY = 
+CLOUDINARY_CONFIG =
